Misc minor cleanups
